Test database 1 and 2
test dtabase 1:

Last week on 9/22/19, I created a database with a server on it. The connectors were called mysql-connector and mysql-connector-python. These connectors were used to write the code in python and make the changes in mysql. I learned that a database is a collection of tables, with various data input. My goal was to create a database using python in mysql with example data. This way, when getting the real data i can be ready as to how i can implement my database. In python i created the database, and changed the data tables with commands. The previous post shows what i was trying to implement. But, il explain a little bit about it. I was trying to make a table name fake_scoring, which had columns website_source, wscore, article_subject, a_score, and final_score. The w_score was based on the the reputability of the website_source, and similarly the a_score, is based on the score of the words in the article_subject of itself. Then the final score is calculated with the adding both the w_score and the a_score. There are a lot of commands that i did to test the table, but the main commands for this code to run is Command 1, Command 2, Command 3, and Command 4. When running the code you have to delete the comment in command 1 and then run it. After that put the commands and again for command 1 and go do the same thing for command 2. Do that for all the four commands, and it should work. Also when creating the server it has to have the same password, user, and host run, as the one in python. Or you make a server with your own password, user, and host, and then change those entries in python.

test database 2:

Then I created a and  new version of the table yesterday. The table included just the the website source, title, data, authoor, and the sources, and grammar. Then I inserted made up data as a test to plan off of when i actually get the actual data into the database. The commands that were used were the five commands listed in test database 2.